backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
